- API 변경에 따른 Response Data Set 변경

+++ app/src/main/java/com/an/howtowear/network/APIService.java
@@ -12,5 +12,5 @@ public interface APIService {
     Call<ForecastListResponse> getForecastList(
             @Query("lat") double lat,
             @Query("lon") double lon,
-            @Query("cnt") double cnt);
+            @Query("cnt") int cnt);
 }
